+++ geo/CHANGES.md
@@ -4,7 +4,7 @@
 
 * Add `LinesIter` algorithm to iterate over the lines in geometries.
   * Very similar to `CoordsIter`, but only implemented where it makes sense (e.g., for `Polygon`, `Rect`, but not `Point`).
-  * TODO: PR ref TBD
+  * <https://github.com/georust/geo/pull/757>
 
 ## 0.19.0